ctf_simulation.py
```python
import json
import requests
from dynamic_code_exec import execute_dynamic_code
from polymorphic_code_gen import generate_polymorphic_code
from api_interaction import send_api_request, receive_api_response
from utils.helpers import parse_data, validate_input
from utils.logging_utils import log_action, log_error
from data.configurations import api_key

class CTFSimulation:
    def __init__(self):
        self.api_key = api_key
        self.session = requests.Session()
        self.session.headers.update({'Authorization': f'Bearer {self.api_key}'})

    def simulate_ctf(self, target_info, scenario):
        try:
            # Validate the input parameters
            validate_input(target_info)
            validate_input(scenario)

            # Log the start of the simulation
            log_action(f"Starting CTF simulation for target: {target_info}")

            # Generate polymorphic code to avoid detection
            polymorphic_code = generate_polymorphic_code()

            # Execute the polymorphic code
            execute_dynamic_code(polymorphic_code)

            # Interact with the AI API to determine the best course of action
            response = send_api_request(self.session, scenario)
            ai_recommendations = receive_api_response(response)

            # Execute AI recommended actions
            for action in ai_recommendations.get('actions', []):
                execute_dynamic_code(action['code'])

            # Log the successful execution of the simulation
            log_action(f"CTF simulation completed for target: {target_info}")

            return True
        except Exception as e:
            log_error(f"Error during CTF simulation: {str(e)}")
            return False

    def evaluate_performance(self, logs):
        try:
            # Analyze the logs to determine the effectiveness of the simulation
            forensic_data = parse_data(logs, 'forensic_data_schema')
            # Further analysis can be added here

            # Log the evaluation results
            log_action("CTF simulation performance evaluation completed.")

            # Generate a post-mortem report
            report = self.generate_post_mortem_report(forensic_data)
            return report
        except Exception as e:
            log_error(f"Error during performance evaluation: {str(e)}")
            return None

    def generate_post_mortem_report(self, forensic_data):
        # This function will generate a detailed report based on the forensic data
        # Placeholder for the actual implementation
        report = {
            "summary": "CTF Simulation Post-Mortem Report",
            "details": forensic_data,
            "conclusion": "Further improvements are suggested."
        }
        return json.dumps(report, indent=4)

# Example usage
if __name__ == "__main__":
    ctf_simulator = CTFSimulation()
    target = {"ip": "192.168.1.1", "port": 8080}
    scenario = {"objective": "Capture the Flag", "difficulty": "High"}
    if ctf_simulator.simulate_ctf(target, scenario):
        logs = "Sample logs from the CTF simulation"
        report = ctf_simulator.evaluate_performance(logs)
        print(report)
```